Edited Checkout Page, Broke down cart component even more
