+++ src/components/ThankYou/ThankYou.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ThankYou = () => {
+  return(
+    <div className="container">
+      <h1>Your order is on its way!</h1>
+    </div>
+  )
+};
+
+export default ThankYou;
